Move coordinate system conversions out of XRInputHelperComponent.
