+++ Assets/Scripts/CoordinateSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f7131bad95afc2a8aed68855e3c8277
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
